@@ -3,46 +3,48 @@
     "version": "0.1.0",
     "private": true,
     "dependencies": {
+        "@ambiot/material-ui-multiple-dates-picker": "^1.0.3",
         "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-        "@emotion/react": "^11.7.0",
-        "@emotion/styled": "^11.6.0",
-        "@fortawesome/fontawesome-svg-core": "^6.2.1",
-        "@fortawesome/free-brands-svg-icons": "^5.15.4",
-        "@fortawesome/free-solid-svg-icons": "^5.15.4",
-        "@fortawesome/react-fontawesome": "^0.1.16",
-        "@googlemaps/markerclusterer": "^2.4.1",
+        "@emotion/react": "^11.11.3",
+        "@emotion/styled": "^11.11.0",
+        "@fortawesome/fontawesome-svg-core": "^6.5.1",
+        "@fortawesome/free-brands-svg-icons": "^6.5.1",
+        "@fortawesome/free-solid-svg-icons": "^6.5.1",
+        "@fortawesome/react-fontawesome": "^0.2.0",
+        "@googlemaps/markerclusterer": "^2.5.3",
         "@googlemaps/react-wrapper": "^1.1.35",
         "@mui/base": "^5.0.0-beta.36",
-        "@mui/core": "^5.0.0-alpha.54",
-        "@mui/icons-material": "^5.2.1",
-        "@mui/lab": "^5.0.0-alpha.73",
-        "@mui/material": "^5.10.6",
-        "@mui/styles": "^5.2.3",
-        "@mui/x-data-grid": "^5.6.0",
-        "@randex/material-ui-multiple-dates-picker": "^0.4.0",
-        "@testing-library/jest-dom": "^4.2.4",
-        "@testing-library/react": "^9.5.0",
-        "@testing-library/user-event": "^7.2.1",
-        "axios": "^0.25.0",
-        "cloudinary-react": "^1.7.1",
-        "moment": "^2.29.0",
+        "@mui/icons-material": "^5.15.10",
+        "@mui/lab": "^5.0.0-alpha.165",
+        "@mui/material": "^5.15.10",
+        "@mui/system": "^5.15.9",
+        "@mui/x-data-grid": "^6.19.4",
+        "@mui/x-date-pickers": "^6.19.4",
+        "@redux-devtools/extension": "^3.3.0",
+        "@testing-library/jest-dom": "^6.4.2",
+        "@testing-library/react": "^14.2.1",
+        "@testing-library/user-event": "^14.5.2",
+        "axios": "^1.6.7",
+        "cloudinary-react": "^1.8.1",
+        "moment": "^2.30.1",
         "prepend-http": "^4.0.0",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
-        "react-moment": "^1.0.0",
-        "react-phone-input-mui": "^2.9.4",
-        "react-player": "^2.9.0",
-        "react-redux": "^7.2.1",
-        "react-router-dom": "^5.2.0",
-        "react-scripts": "^5.0.0",
-        "react-spring": "^9.3.2",
-        "react-use": "^17.3.2",
-        "redux": "^4.0.5",
-        "redux-devtools-extension": "^2.13.8",
-        "redux-thunk": "^2.3.0",
-        "spinners-react": "^1.0.6",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-moment": "^1.1.3",
+        "react-phone-input-2": "^2.15.1",
+        "react-player": "^2.14.1",
+        "react-redux": "^9.1.0",
+        "react-router-dom": "^6.22.0",
+        "react-spring": "^9.7.3",
+        "react-use": "^17.5.0",
+        "redux": "^5.0.1",
+        "redux-thunk": "^3.1.0",
+        "spinners-react": "^1.0.7",
         "us-state-converter": "^1.0.6",
         "uuid": "^9.0.1"
+    },
+    "devDependencies": {
+        "react-scripts": "^5.0.1"
     },
     "scripts": {
         "start": "react-scripts start",
