@@ -1,9 +1,9 @@
 import React, { Fragment, useEffect } from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import Navbar from './components/layout/Navbar';
 import SwipeableDrawer from './components/layout/SwipeableDrawer';
 //import Landing from './components/layout/Landing';
-import Routes from './components/routing/Routes';
+import TheRoutes from './components/routing/Routes';
 
 // Redux
 import { Provider } from 'react-redux';
@@ -11,7 +11,12 @@ import store from './store';
 import { loadUser } from './actions/auth';
 import setAuthToken from './utils/setAuthToken';
 
-import { createTheme, ThemeProvider } from '@mui/material/styles';
+import {
+    createTheme,
+    ThemeProvider,
+    StyledEngineProvider,
+    // adaptV4Theme,
+} from '@mui/material/styles';
 
 //Consider checking out material-ui and https://material-kit-pro-react.devias.io/
 import './App.css';
@@ -61,18 +66,20 @@ const App = () => {
     }, []);
     return (
         <Provider store={store}>
-            <ThemeProvider theme={darkTheme}>
-                <Router>
-                    <Fragment>
-                        <Navbar />
-                        <SwipeableDrawer className="porchlightBG" />
-                        <Switch>
-                            {/* <Route exact path='/' component={Landing} /> */}
-                            <Route component={Routes} />
-                        </Switch>
-                    </Fragment>
-                </Router>
-            </ThemeProvider>
+            <StyledEngineProvider injectFirst>
+                <ThemeProvider theme={darkTheme}>
+                    <Router>
+                        <Fragment>
+                            <Navbar />
+                            <SwipeableDrawer className="porchlightBG" />
+                            <TheRoutes />
+                            {/* <Routes>
+                                <Route component={TheRoutes} />
+                            </Routes> */}
+                        </Fragment>
+                    </Router>
+                </ThemeProvider>
+            </StyledEngineProvider>
         </Provider>
     );
 };
