@@ -1,6 +1,10 @@
 import axios from 'axios'; //only for uploads as of December 31st, 2021
 import React, { Fragment, useState, useEffect, useRef } from 'react';
-import { Link, withRouter, useLocation } from 'react-router-dom';
+import {
+    Link,
+    // withRouter,
+    useLocation,
+} from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { connect, useDispatch } from 'react-redux';
 import { IMAGE_UPLOAD, UPDATE_ARTIST_ME } from '../../actions/types';
@@ -31,7 +35,7 @@ import {
     withStyles,
     Typography,
 } from '@mui/material';
-import ReactPhoneInput from 'react-phone-input-mui';
+import { PhoneInput as ReactPhoneInput } from 'react-phone-input-2';
 import { styled } from '@mui/material/styles';
 import Button from '../layout/SvgButton';
 
@@ -46,7 +50,7 @@ import AccountBoxTwoToneIcon from '@mui/icons-material/AccountBoxTwoTone';
 
 //import { DateRangePicker, DateRange } from "materialui-daterange-picker";
 //import MultipleDatesPicker from '@randex/material-ui-multiple-dates-picker';
-import MultipleDatesPicker from '../mui-multi-date-picker-lib';
+// import MultipleDatesPicker from '../mui-multi-date-picker-lib';
 
 import { useTransition, animated, config } from '@react-spring/web';
 import styles from '../../formCards.css';
@@ -1195,6 +1199,7 @@ const EditArtistProfileForm = ({
                                               idx
                                           )
                                       }
+                                      size="large"
                                   >
                                       <DeleteIcon />
                                   </IconButton>
@@ -1898,4 +1903,5 @@ export default connect(mapStateToProps, {
     createMyArtist,
     getArtists,
     updateUserAvatar,
-})(withRouter(EditArtistProfileForm)); //withRouter allows us to pass history objects
+    // })(withRouter(EditArtistProfileForm)); //withRouter allows us to pass history objects
+})(EditArtistProfileForm); //withRouter allows us to pass history objects
