@@ -1,7 +1,7 @@
 import React, { Fragment, useState, useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
-import { Link, withRouter } from 'react-router-dom';
+// import { Link, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import { Grid, Typography, FormGroup, Tooltip } from '@mui/material';
@@ -146,4 +146,4 @@ const mapStateToProps = (state) => ({
     myArtistEvents: state.event.myArtistEvents,
 });
 
-export default connect(mapStateToProps, {})(withRouter(PastArtistEvents)); //withRouter allows us to pass history objects
+export default connect(mapStateToProps, {})(PastArtistEvents);
