@@ -1,6 +1,10 @@
 //originally copied EditArtistBookingForm.js
 import React, { Fragment, useState, useEffect, useRef } from 'react';
-import { Link, withRouter, useLocation } from 'react-router-dom';
+import {
+    Link,
+    // withRouter,
+    useLocation,
+} from 'react-router-dom';
 import PropTypes from 'prop-types';
 import { connect, useDispatch } from 'react-redux';
 import { IMAGE_UPLOAD, UPDATE_ARTIST_ME } from '../../actions/types';
@@ -44,7 +48,8 @@ import DateRangeTwoToneIcon from '@mui/icons-material/DateRangeTwoTone';
 
 //import { DateRangePicker, DateRange } from "materialui-daterange-picker";
 //import MultipleDatesPicker from '@randex/material-ui-multiple-dates-picker';
-import MultipleDatesPicker from '../mui-multi-date-picker-lib';
+// import MultipleDatesPicker from '../mui-multi-date-picker-lib';
+import { DateCalendar } from '@mui/x-date-pickers';
 
 import { useTransition, animated, config } from '@react-spring/web';
 import styles from '../../formCards.css';
@@ -629,7 +634,33 @@ const ArtistEventForm = ({
                 // 	  ))
                 // 	: '',
 
-                <MultipleDatesPicker
+                // <MultipleDatesPicker
+                //     id="bookingWhen"
+                //     name="bookingWhen"
+                //     open={true}
+                //     //trying to figure out how to disable dates you've already picked ~Aug 26, 2022
+                //     disabledDates={
+                //         myArtistEvents &&
+                //         myArtistEvents
+                //             .filter((event) => {
+                //                 if (event.bookingWhen !== bookingWhen) {
+                //                     return true;
+                //                 } else {
+                //                     return false;
+                //                 }
+                //             })
+                //             .map((event) => {
+                //                 return event.bookingWhen;
+                //             })
+                //     }
+                //     //readOnly={bookingWhen ? true : false} //if there's a bookingWhen date, don't let people change it
+                //     selectedDates={bookingWhen ? [bookingWhen] : []}
+                //     //value={bookingWhen}
+                //     onCancel={() => setOpen(false)}
+                //     //onSubmit={dates => console.log('selected dates', dates)}
+                //     onChange={(target) => onCalendarChange(target)}
+                // />,
+                <DateCalendar
                     id="bookingWhen"
                     name="bookingWhen"
                     open={true}
@@ -1412,6 +1443,7 @@ const ArtistEventForm = ({
                                               idx
                                           )
                                       }
+                                      size="large"
                                   >
                                       <DeleteIcon />
                                   </IconButton>
@@ -2189,4 +2221,5 @@ export default connect(mapStateToProps, {
     jumpTo,
     getHostsLocations,
     editArtistEvent,
-})(withRouter(ArtistEventForm)); //withRouter allows us to pass history objects
+    // })(withRouter(ArtistEventForm)); //withRouter allows us to pass history objects
+})(ArtistEventForm); //withRouter allows us to pass history objects
