@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { Link, withRouter, useHistory } from 'react-router-dom';
+// import { Link, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import Button from '../layout/SvgButton';
@@ -114,5 +114,6 @@ const mapStateToProps = (state) => ({
 
 //export default EditHostEvent;
 export default connect(mapStateToProps, { jumpTo, closeEventEditDrawer })(
-    withRouter(EditHostEvent)
+    // withRouter(EditHostEvent)
+    EditHostEvent
 ); //withRouter allows us to pass history objects
