@@ -1,7 +1,7 @@
 import React, { Fragment, useState, useRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
 
-import { Link, withRouter } from 'react-router-dom';
+// import { Link, withRouter } from 'react-router-dom';
 import { connect } from 'react-redux';
 
 import {
@@ -110,4 +110,5 @@ const mapStateToProps = (state) => ({
 });
 
 //export default HostAdminActiveFalse;
-export default connect(mapStateToProps, {})(withRouter(HostAdminActiveFalse)); //withRouter allows us to pass history objects
+// export default connect(mapStateToProps, {})(withRouter(HostAdminActiveFalse)); //withRouter allows us to pass history objects
+export default connect(mapStateToProps, {})(HostAdminActiveFalse); //withRouter allows us to pass history objects
