@@ -14,7 +14,7 @@ const styles = theme => ({
   },
   day: {
     margin: `0 ${theme.spacing(1)}`,
-    [theme.breakpoints.down('md')]: {
+    [theme.breakpoints.down('xl')]: {
       margin: `0 2px`,
     }
   },
@@ -22,7 +22,7 @@ const styles = theme => ({
     width: 36,
     height: 36,
     margin: `0 ${theme.spacing(1)}`,
-    [theme.breakpoints.down('md')]: {
+    [theme.breakpoints.down('xl')]: {
       margin: `0 2px`,
     }
   }
