@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import thunk from 'redux-thunk';
+import { composeWithDevTools } from '@redux-devtools/extension';
+import { thunk } from 'redux-thunk';
 import rootReducer from './reducers'; // ./reducers/index.js
 
 const initialState = {};
@@ -8,9 +8,9 @@ const initialState = {};
 const middleware = [thunk];
 
 const store = createStore(
-  rootReducer,
-  initialState,
-  composeWithDevTools(applyMiddleware(...middleware))
+    rootReducer,
+    initialState,
+    composeWithDevTools(applyMiddleware(...middleware))
 );
 
 export default store;
